Subiendo AppointmentDTO y Exception

+++ src/main/java/com/dh/DentalClinicMVC/service/impl/DentistServiceImpl.java
@@ -1,6 +1,7 @@
 package com.dh.DentalClinicMVC.service.impl;
 
-import com.dh.DentalClinicMVC.model.Dentist;
+import com.dh.DentalClinicMVC.entity.Dentist;
+import com.dh.DentalClinicMVC.exception.ResourceNotFoundException;
 import com.dh.DentalClinicMVC.repository.DentistRepository;
 import com.dh.DentalClinicMVC.service.IDentistService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -35,12 +36,29 @@ public class DentistServiceImpl implements IDentistService {
     }
 
     @Override
-    public void delete(Long id) {
-        dentistRepository.deleteById(id);
+    public void delete(Long id) throws ResourceNotFoundException{
+        //Buscamos por id el dentist
+        // y si no existe vamos a lanzar una exception
+
+        Optional<Dentist> dentist = findById(id);
+
+        if (dentist.isPresent()){
+            dentistRepository.deleteById(id);
+        }else{
+            throw new ResourceNotFoundException("No se puddo eliminar el Odontologo con ID: " +
+                    id + ", no se encuentra en la BD");
+        }
+
     }
 
     @Override
     public List<Dentist> findAll() {
         return dentistRepository.findAll();
     }
+
+    @Override
+    public Optional<Dentist> findByRegistration(Integer registration) {
+        return dentistRepository.findByRegistration(registration);
+    }
+
 }
